refactor: split interface and bucket
