+++ src/general_manager/interface/__init__.py
@@ -1,3 +1,4 @@
-# from .baseInterface import InterfaceBase, Bucket
 # from .databaseInterface import DatabaseInterface, ReadOnlyInterface, DatabaseBucket
 # from .calculationInterface import CalculationInterface, CalculationBucket
+# from .readOnlyInterface import ReadOnlyInterface
+# from .databaseInterface import DatabaseInterface
